Fix receipt timeout bug
Fix the receipt timeout bug where the timer would decrease by multiple
seconds each second, depending how many times the receipt screen had
been accessed. This was because each time the timer was triggered, a
separate instance of calling the 'update_countdown()' function was bound
to the countdown signal. This commit moves the signal binding to the
screen constructor, so it is run just once and resolves this problem