Rewrite mqtt to be a class abstraction
Rewrite the mqtt implementation to be a class which is an abstraction of
the mqtt client library we are using. This class is meant to provide a
very simple interface for the controller, where it can just provide a
topic name and a callback function for messages on that topic.